prevent minor bug issue

+++ src/les/module/ewald.py
@@ -49,7 +49,7 @@ class Ewald(nn.Module):
         results: List[torch.Tensor] = []
         q_induced_results: List[torch.Tensor] = []
         u_induced_results: List[torch.Tensor] = []
-        for i in unique_batches:
+        for i in unique_batches.long():
             mask = batch == i  # Create a mask for the i-th configuration
             # Calculate the potential energy for the i-th configuration
             r_raw_now, q_now = r[mask], q[mask]
